+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Unreleased
 ----------
 
 * Visualization - resolved a bug where a schedule was modified when drawing a circuit diagram (#197, !250)
+* Visualization - made matplotlib schedule visualization methods accessible as methods `plot_circuit_diagram_mpl` and `plot_pulse_diagram_mpl` of the `Schedule` class (!253)
 
 Breaking changes
 ~~~~~~~~~~~~~~~~